core: decouple security handlers with different revisions from PdfCrypt
