+++ pdf/core/crypt.go
@@ -6,19 +6,10 @@
 package core
 
 import (
-	"bytes"
-	"crypto/aes"
-	"crypto/cipher"
 	"crypto/md5"
 	"crypto/rand"
-	"crypto/rc4"
-	"crypto/sha256"
-	"crypto/sha512"
-	"encoding/binary"
 	"errors"
 	"fmt"
-	"hash"
-	"io"
 	"math"
 	"time"
 
@@ -546,47 +537,30 @@ func (crypt *PdfCrypt) GetAccessPermissions() AccessPermissions {
 	return crypt.encryptStd.P
 }
 
-// Check whether the specified password can be used to decrypt the document.
-func (crypt *PdfCrypt) authenticate(password []byte) (bool, error) {
-	// Also build the encryption/decryption key.
-
-	crypt.authenticated = false
+func (crypt *PdfCrypt) securityHandler() stdSecurityHandler {
 	if crypt.encryptStd.R >= 5 {
-		authenticated, err := crypt.alg2a(password)
-		if err != nil {
-			return false, err
-		}
-		crypt.authenticated = authenticated
-		return authenticated, err
+		return stdHandlerR6{}
 	}
+	return stdHandlerR4{
+		ID0:    crypt.id0,
+		Length: crypt.encrypt.Length,
+	}
+}
 
-	// Try user password.
-	common.Log.Trace("Debugging authentication - user pass")
-	authenticated, err := crypt.alg6(password)
+// Check whether the specified password can be used to decrypt the document.
+// Also build the encryption/decryption key.
+func (crypt *PdfCrypt) authenticate(password []byte) (bool, error) {
+	crypt.authenticated = false
+	h := crypt.securityHandler()
+	fkey, perm, err := h.Authenticate(&crypt.encryptStd, password)
 	if err != nil {
 		return false, err
+	} else if perm == 0 || len(fkey) == 0 {
+		return false, nil
 	}
-	if authenticated {
-		common.Log.Trace("this.authenticated = True")
-		crypt.authenticated = true
-		return true, nil
-	}
-
-	// Try owner password also.
-	// May not be necessary if only want to get all contents.
-	// (user pass needs to be known or empty).
-	common.Log.Trace("Debugging authentication - owner pass")
-	authenticated, err = crypt.alg7(password)
-	if err != nil {
-		return false, err
-	}
-	if authenticated {
-		common.Log.Trace("this.authenticated = True")
-		crypt.authenticated = true
-		return true, nil
-	}
-
-	return false, nil
+	crypt.authenticated = true
+	crypt.encryptionKey = fkey
+	return true, nil
 }
 
 // Check access rights and permissions for a specified password.  If either user/owner password is specified,
@@ -596,68 +570,15 @@ func (crypt *PdfCrypt) authenticate(password []byte) (bool, error) {
 // The AccessPermissions shows what access the user has for editing etc.
 // An error is returned if there was a problem performing the authentication.
 func (crypt *PdfCrypt) checkAccessRights(password []byte) (bool, AccessPermissions, error) {
-	// Try owner password -> full rights.
-	var (
-		isOwner bool
-		err     error
-	)
-	if crypt.encryptStd.R >= 5 {
-		var h []byte
-		h, err = crypt.alg12(password)
-		if err != nil {
-			return false, 0, err
-		}
-		isOwner = len(h) != 0
-	} else {
-		isOwner, err = crypt.alg7(password)
-	}
+	h := crypt.securityHandler()
+	// TODO(dennwc): it computes an encryption key as well; if necessary, define a new interface method to optimize this
+	fkey, perm, err := h.Authenticate(&crypt.encryptStd, password)
 	if err != nil {
 		return false, 0, err
+	} else if perm == 0 || len(fkey) == 0 {
+		return false, 0, nil
 	}
-	if isOwner {
-		// owner -> full rights.
-		return true, PermOwner, nil
-	}
-
-	// Try user password.
-	var isUser bool
-	if crypt.encryptStd.R >= 5 {
-		var h []byte
-		h, err = crypt.alg11(password)
-		if err != nil {
-			return false, 0, err
-		}
-		isUser = len(h) != 0
-	} else {
-		isUser, err = crypt.alg6(password)
-	}
-	if err != nil {
-		return false, 0, err
-	}
-	if isUser {
-		// User password specified correctly -> access granted with specified permissions.
-		return true, crypt.encryptStd.P, nil
-	}
-
-	// Cannot even view the file.
-	return false, 0, nil
-}
-
-func (crypt *PdfCrypt) paddedPass(pass []byte) []byte {
-	key := make([]byte, 32)
-	if len(pass) >= 32 {
-		for i := 0; i < 32; i++ {
-			key[i] = pass[i]
-		}
-	} else {
-		for i := 0; i < len(pass); i++ {
-			key[i] = pass[i]
-		}
-		for i := len(pass); i < 32; i++ {
-			key[i] = padding[i-len(pass)]
-		}
-	}
-	return key
+	return true, perm, nil
 }
 
 // Generates a key for encrypting a specific object based on the
@@ -1073,724 +994,13 @@ func (crypt *PdfCrypt) Encrypt(obj PdfObject, parentObjNum, parentGenNum int64) 
 	return nil
 }
 
-// aesZeroIV allocates a zero-filled buffer that serves as an initialization vector for AESv3.
-func (crypt *PdfCrypt) aesZeroIV() []byte {
-	if crypt.ivAESZero == nil {
-		crypt.ivAESZero = make([]byte, aes.BlockSize)
-	}
-	return crypt.ivAESZero
-}
-
-// alg2a retrieves the encryption key from an encrypted document (R >= 5).
-// It returns false if the password was wrong.
-// 7.6.4.3.2 Algorithm 2.A (page 83)
-func (crypt *PdfCrypt) alg2a(pass []byte) (bool, error) {
-	// O & U: 32 byte hash + 8 byte Validation Salt + 8 byte Key Salt
-
-	// step a: Unicode normalization
-	// TODO(dennwc): make sure that UTF-8 strings are normalized
-
-	// step b: truncate to 127 bytes
-	if len(pass) > 127 {
-		pass = pass[:127]
-	}
-
-	// step c: test pass against the owner key
-	h, err := crypt.alg12(pass)
-	if err != nil {
-		return false, err
-	}
-	var (
-		data []byte // data to hash
-		ekey []byte // encrypted file key
-		ukey []byte // user key; set only when using owner's password
-	)
-	if len(h) != 0 {
-		// owner password valid
-
-		// step d: compute an intermediate owner key
-		str := make([]byte, len(pass)+8+48)
-		i := copy(str, pass)
-		i += copy(str[i:], crypt.encryptStd.O[40:48]) // owner Key Salt
-		i += copy(str[i:], crypt.encryptStd.U[0:48])
-
-		data = str
-		ekey = crypt.encryptStd.OE
-		ukey = crypt.encryptStd.U[0:48]
-	} else {
-		// check user password
-		h, err = crypt.alg11(pass)
-		if err == nil && len(h) == 0 {
-			// try default password
-			h, err = crypt.alg11([]byte(""))
-		}
-		if err != nil {
-			return false, err
-		} else if len(h) == 0 {
-			// wrong password
-			return false, nil
-		}
-		// step e: compute an intermediate user key
-		str := make([]byte, len(pass)+8)
-		i := copy(str, pass)
-		i += copy(str[i:], crypt.encryptStd.U[40:48]) // user Key Salt
-
-		data = str
-		ekey = crypt.encryptStd.UE
-		ukey = nil
-	}
-	ekey = ekey[:32]
-
-	// intermediate key
-	ikey := crypt.alg2b(data, pass, ukey)
-
-	ac, err := aes.NewCipher(ikey[:32])
-	if err != nil {
-		panic(err)
-	}
-
-	iv := crypt.aesZeroIV()
-	cbc := cipher.NewCBCDecrypter(ac, iv)
-	fkey := make([]byte, 32)
-	cbc.CryptBlocks(fkey, ekey)
-
-	crypt.encryptionKey = fkey
-
-	if crypt.encryptStd.R == 5 {
-		return true, nil
-	}
-
-	return crypt.alg13(fkey)
-}
-
-// alg2b computes a hash for R=5 and R=6.
-func (crypt *PdfCrypt) alg2b(data, pwd, userKey []byte) []byte {
-	if crypt.encryptStd.R == 5 {
-		return alg2b_R5(data)
-	}
-	return alg2b(data, pwd, userKey)
-}
-
-// alg2b_R5 computes a hash for R=5, used in a deprecated extension.
-// It's used the same way as a hash described in Algorithm 2.B, but it doesn't use the original password
-// and the user key to calculate the hash.
-func alg2b_R5(data []byte) []byte {
-	h := sha256.New()
-	h.Write(data)
-	return h.Sum(nil)
-}
-
-// repeat repeats first n bytes of buf until the end of the buffer.
-// It assumes that the length of buf is a multiple of n.
-func repeat(buf []byte, n int) {
-	bp := n
-	for bp < len(buf) {
-		copy(buf[bp:], buf[:bp])
-		bp *= 2
-	}
-}
-
-// alg2b computes a hash for R=6.
-// 7.6.4.3.3 Algorithm 2.B (page 83)
-func alg2b(data, pwd, userKey []byte) []byte {
-	var (
-		s256, s384, s512 hash.Hash
-	)
-	s256 = sha256.New()
-	hbuf := make([]byte, 64)
-
-	h := s256
-	h.Write(data)
-	K := h.Sum(hbuf[:0])
-
-	buf := make([]byte, 64*(127+64+48))
-
-	round := func(rnd int) (E []byte) {
-		// step a: repeat pass+K 64 times
-		n := len(pwd) + len(K) + len(userKey)
-		part := buf[:n]
-		i := copy(part, pwd)
-		i += copy(part[i:], K[:])
-		i += copy(part[i:], userKey)
-		if i != n {
-			panic("wrong size")
-		}
-		K1 := buf[:n*64]
-		repeat(K1, n)
-
-		// step b: encrypt K1 with AES-128 CBC
-		ac, err := aes.NewCipher(K[0:16])
-		if err != nil {
-			panic(err)
-		}
-		cbc := cipher.NewCBCEncrypter(ac, K[16:32])
-		cbc.CryptBlocks(K1, K1)
-		E = K1
-
-		// step c: use 16 bytes of E as big-endian int, select the next hash
-		b := 0
-		for i := 0; i < 16; i++ {
-			b += int(E[i] % 3)
-		}
-		var h hash.Hash
-		switch b % 3 {
-		case 0:
-			h = s256
-		case 1:
-			if s384 == nil {
-				s384 = sha512.New384()
-			}
-			h = s384
-		case 2:
-			if s512 == nil {
-				s512 = sha512.New()
-			}
-			h = s512
-		}
-
-		// step d: take the hash of E, use as a new K
-		h.Reset()
-		h.Write(E)
-		K = h.Sum(hbuf[:0])
-
-		return E
-	}
-
-	for i := 0; ; {
-		E := round(i)
-		b := uint8(E[len(E)-1])
-		// from the spec, it appears that i should be incremented after
-		// the test, but that doesn't match what Adobe does
-		i++
-		if i >= 64 && b <= uint8(i-32) {
-			break
-		}
-	}
-	return K[:32]
-}
-
-// alg2 computes an encryption key.
-func (crypt *PdfCrypt) alg2(pass []byte) []byte {
-	common.Log.Trace("alg2")
-	key := crypt.paddedPass(pass)
-
-	h := md5.New()
-	h.Write(key)
-
-	// Pass O.
-	h.Write(crypt.encryptStd.O)
-
-	// Pass P (Lower order byte first).
-	var p = uint32(crypt.encryptStd.P)
-	var pb []byte
-	for i := 0; i < 4; i++ {
-		pb = append(pb, byte(((p >> uint(8*i)) & 0xff)))
-	}
-	h.Write(pb)
-	common.Log.Trace("go P: % x", pb)
-
-	// Pass ID[0] from the trailer
-	h.Write([]byte(crypt.id0))
-
-	common.Log.Trace("this.R = %d encryptMetadata %v", crypt.encryptStd.R, crypt.encryptStd.EncryptMetadata)
-	if (crypt.encryptStd.R >= 4) && !crypt.encryptStd.EncryptMetadata {
-		h.Write([]byte{0xff, 0xff, 0xff, 0xff})
-	}
-	hashb := h.Sum(nil)
-
-	if crypt.encryptStd.R >= 3 {
-		for i := 0; i < 50; i++ {
-			h = md5.New()
-			h.Write(hashb[0 : crypt.encrypt.Length/8])
-			hashb = h.Sum(nil)
-		}
-	}
-
-	if crypt.encryptStd.R >= 3 {
-		return hashb[0 : crypt.encrypt.Length/8]
-	}
-
-	return hashb[0:5]
-}
-
-// Create the RC4 encryption key.
-func (crypt *PdfCrypt) alg3Key(pass []byte) []byte {
-	h := md5.New()
-	okey := crypt.paddedPass(pass)
-	h.Write(okey)
-
-	if crypt.encryptStd.R >= 3 {
-		for i := 0; i < 50; i++ {
-			hashb := h.Sum(nil)
-			h = md5.New()
-			h.Write(hashb)
-		}
-	}
-
-	encKey := h.Sum(nil)
-	if crypt.encryptStd.R == 2 {
-		encKey = encKey[0:5]
-	} else {
-		encKey = encKey[0 : crypt.encrypt.Length/8]
-	}
-	return encKey
-}
-
-// alg3 computes the encryption dictionary’s O (owner password) value.
-func (crypt *PdfCrypt) alg3(upass, opass []byte) (string, error) {
-	// Return O string val.
-	O := ""
-
-	var encKey []byte
-	if len(opass) > 0 {
-		encKey = crypt.alg3Key(opass)
-	} else {
-		encKey = crypt.alg3Key(upass)
-	}
-
-	ociph, err := rc4.NewCipher(encKey)
-	if err != nil {
-		return O, errors.New("Failed rc4 ciph")
-	}
-
-	ukey := crypt.paddedPass(upass)
-	encrypted := make([]byte, len(ukey))
-	ociph.XORKeyStream(encrypted, ukey)
-
-	if crypt.encryptStd.R >= 3 {
-		encKey2 := make([]byte, len(encKey))
-		for i := 0; i < 19; i++ {
-			for j := 0; j < len(encKey); j++ {
-				encKey2[j] = encKey[j] ^ byte(i+1)
-			}
-			ciph, err := rc4.NewCipher(encKey2)
-			if err != nil {
-				return O, errors.New("Failed rc4 ciph")
-			}
-			ciph.XORKeyStream(encrypted, encrypted)
-		}
-	}
-
-	O = string(encrypted)
-	return O, nil
-}
-
-// alg4 computes the encryption dictionary’s U (user password) value (Security handlers of revision 2).
-func (crypt *PdfCrypt) alg4(upass []byte) (string, []byte, error) {
-	U := ""
-
-	ekey := crypt.alg2(upass)
-	ciph, err := rc4.NewCipher(ekey)
-	if err != nil {
-		return U, ekey, errors.New("Failed rc4 ciph")
-	}
-
-	s := []byte(padding)
-	encrypted := make([]byte, len(s))
-	ciph.XORKeyStream(encrypted, s)
-
-	U = string(encrypted)
-	return U, ekey, nil
-}
-
-// alg5 computes the encryption dictionary’s U (user password) value (Security handlers of revision 3 or greater).
-func (crypt *PdfCrypt) alg5(upass []byte) (string, []byte, error) {
-	U := ""
-
-	ekey := crypt.alg2(upass)
-
-	h := md5.New()
-	h.Write([]byte(padding))
-	h.Write([]byte(crypt.id0))
-	hash := h.Sum(nil)
-
-	common.Log.Trace("alg5")
-	common.Log.Trace("ekey: % x", ekey)
-	common.Log.Trace("ID: % x", crypt.id0)
-
-	if len(hash) != 16 {
-		return U, ekey, errors.New("Hash length not 16 bytes")
-	}
-
-	ciph, err := rc4.NewCipher(ekey)
-	if err != nil {
-		return U, ekey, errors.New("Failed rc4 ciph")
-	}
-	encrypted := make([]byte, 16)
-	ciph.XORKeyStream(encrypted, hash)
-
-	// Do the following 19 times: Take the output from the previous
-	// invocation of the RC4 function and pass it as input to a new
-	// invocation of the function; use an encryption key generated by
-	// taking each byte of the original encryption key obtained in step
-	// (a) and performing an XOR (exclusive or) operation between that
-	// byte and the single-byte value of the iteration counter (from 1 to 19).
-	ekey2 := make([]byte, len(ekey))
-	for i := 0; i < 19; i++ {
-		for j := 0; j < len(ekey); j++ {
-			ekey2[j] = ekey[j] ^ byte(i+1)
-		}
-		ciph, err = rc4.NewCipher(ekey2)
-		if err != nil {
-			return U, ekey, errors.New("Failed rc4 ciph")
-		}
-		ciph.XORKeyStream(encrypted, encrypted)
-		common.Log.Trace("i = %d, ekey: % x", i, ekey2)
-		common.Log.Trace("i = %d -> % x", i, encrypted)
-	}
-
-	bb := make([]byte, 32)
-	for i := 0; i < 16; i++ {
-		bb[i] = encrypted[i]
-	}
-
-	// Append 16 bytes of arbitrary padding to the output from the final
-	// invocation of the RC4 function and store the 32-byte result as
-	// the value of the U entry in the encryption dictionary.
-	_, err = rand.Read(bb[16:32])
-	if err != nil {
-		return U, ekey, errors.New("Failed to gen rand number")
-	}
-
-	U = string(bb)
-	return U, ekey, nil
-}
-
-// alg6 authenticates the user password.
-func (crypt *PdfCrypt) alg6(upass []byte) (bool, error) {
-	var uo string
-	var err error
-	var key []byte
-	if crypt.encryptStd.R == 2 {
-		uo, key, err = crypt.alg4(upass)
-	} else if crypt.encryptStd.R >= 3 {
-		uo, key, err = crypt.alg5(upass)
-	} else {
-		return false, errors.New("invalid R")
-	}
-
-	if err != nil {
-		return false, err
-	}
-
-	common.Log.Trace("check: % x == % x ?", string(uo), string(crypt.encryptStd.U))
-
-	uGen := string(uo)                 // Generated U from specified pass.
-	uDoc := string(crypt.encryptStd.U) // U from the document.
-	if crypt.encryptStd.R >= 3 {
-		// comparing on the first 16 bytes in the case of security
-		// handlers of revision 3 or greater),
-		if len(uGen) > 16 {
-			uGen = uGen[0:16]
-		}
-		if len(uDoc) > 16 {
-			uDoc = uDoc[0:16]
-		}
-	}
-
-	if uGen == uDoc {
-		crypt.encryptionKey = key
-		return true, nil
-	}
-
-	return false, nil
-}
-
-// alg7 authenticates the owner password.
-func (crypt *PdfCrypt) alg7(opass []byte) (bool, error) {
-	encKey := crypt.alg3Key(opass)
-
-	decrypted := make([]byte, len(crypt.encryptStd.O))
-	if crypt.encryptStd.R == 2 {
-		ciph, err := rc4.NewCipher(encKey)
-		if err != nil {
-			return false, errors.New("Failed cipher")
-		}
-		ciph.XORKeyStream(decrypted, crypt.encryptStd.O)
-	} else if crypt.encryptStd.R >= 3 {
-		s := append([]byte{}, crypt.encryptStd.O...)
-		for i := 0; i < 20; i++ {
-			//newKey := encKey
-			newKey := append([]byte{}, encKey...)
-			for j := 0; j < len(encKey); j++ {
-				newKey[j] ^= byte(19 - i)
-			}
-			ciph, err := rc4.NewCipher(newKey)
-			if err != nil {
-				return false, errors.New("Failed cipher")
-			}
-			ciph.XORKeyStream(decrypted, s)
-			s = append([]byte{}, decrypted...)
-		}
-	} else {
-		return false, errors.New("invalid R")
-	}
-
-	auth, err := crypt.alg6(decrypted)
-	if err != nil {
-		return false, nil
-	}
-
-	return auth, nil
-}
-
 // generateParams generates encryption parameters for specified passwords.
 func (crypt *PdfCrypt) generateParams(upass, opass []byte) error {
-	if crypt.encryptStd.R < 5 {
-		// Make the O and U objects.
-		O, err := crypt.alg3(upass, opass)
-		if err != nil {
-			common.Log.Debug("ERROR: Error generating O for encryption (%s)", err)
-			return err
-		}
-		crypt.encryptStd.O = []byte(O)
-		common.Log.Trace("gen O: % x", O)
-		U, key, err := crypt.alg5(upass)
-		if err != nil {
-			common.Log.Debug("ERROR: Error generating O for encryption (%s)", err)
-			return err
-		}
-		common.Log.Trace("gen U: % x", U)
-		crypt.encryptStd.U = []byte(U)
-		crypt.encryptionKey = key
-		return nil
-	}
-	crypt.encryptionKey = make([]byte, 32)
-	if _, err := io.ReadFull(rand.Reader, crypt.encryptionKey); err != nil {
-		return err
-	}
-	return crypt.generateR6(upass, opass)
-}
-
-// generateR6 is the algorithm opposite to alg2a (R>=5).
-// It generates U,O,UE,OE,Perms fields using AESv3 encryption.
-// There is no algorithm number assigned to this function in the spec.
-func (crypt *PdfCrypt) generateR6(upass, opass []byte) error {
-	// all these field will be populated by functions below
-	crypt.encryptStd.U = nil
-	crypt.encryptStd.O = nil
-	crypt.encryptStd.UE = nil
-	crypt.encryptStd.OE = nil
-	crypt.encryptStd.Perms = nil // populated only for R=6
-
-	if len(upass) > 127 {
-		upass = upass[:127]
-	}
-	if len(opass) > 127 {
-		opass = opass[:127]
-	}
-	// generate U and UE
-	if err := crypt.alg8(upass); err != nil {
-		return err
-	}
-	// generate O and OE
-	if err := crypt.alg9(opass); err != nil {
-		return err
-	}
-	if crypt.encryptStd.R == 5 {
-		return nil
-	}
-	// generate Perms
-	return crypt.alg10()
-}
-
-// alg8 computes the encryption dictionary's U (user password) and UE (user encryption) values (R>=5).
-// 7.6.4.4.6 Algorithm 8 (page 86)
-func (crypt *PdfCrypt) alg8(upass []byte) error {
-	// step a: compute U (user password)
-	var rbuf [16]byte
-	if _, err := io.ReadFull(rand.Reader, rbuf[:]); err != nil {
-		return err
-	}
-	valSalt := rbuf[0:8]
-	keySalt := rbuf[8:16]
-
-	str := make([]byte, len(upass)+len(valSalt))
-	i := copy(str, upass)
-	i += copy(str[i:], valSalt)
-
-	h := crypt.alg2b(str, upass, nil)
-
-	U := make([]byte, len(h)+len(valSalt)+len(keySalt))
-	i = copy(U, h[:32])
-	i += copy(U[i:], valSalt)
-	i += copy(U[i:], keySalt)
-
-	crypt.encryptStd.U = U
-
-	// step b: compute UE (user encryption)
-
-	// str still contains a password, reuse it
-	i = len(upass)
-	i += copy(str[i:], keySalt)
-
-	h = crypt.alg2b(str, upass, nil)
-
-	ac, err := aes.NewCipher(h[:32])
+	h := crypt.securityHandler()
+	ekey, err := h.GenerateParams(&crypt.encryptStd, opass, upass)
 	if err != nil {
-		panic(err)
+		return err
 	}
-
-	iv := crypt.aesZeroIV()
-	cbc := cipher.NewCBCEncrypter(ac, iv)
-	UE := make([]byte, 32)
-	cbc.CryptBlocks(UE, crypt.encryptionKey[:32])
-	crypt.encryptStd.UE = UE
-
+	crypt.encryptionKey = ekey
 	return nil
-}
-
-// alg9 computes the encryption dictionary's O (owner password) and OE (owner encryption) values (R>=5).
-// 7.6.4.4.7 Algorithm 9 (page 86)
-func (crypt *PdfCrypt) alg9(opass []byte) error {
-	// step a: compute O (owner password)
-	var rbuf [16]byte
-	if _, err := io.ReadFull(rand.Reader, rbuf[:]); err != nil {
-		return err
-	}
-	valSalt := rbuf[0:8]
-	keySalt := rbuf[8:16]
-	userKey := crypt.encryptStd.U[:48]
-
-	str := make([]byte, len(opass)+len(valSalt)+len(userKey))
-	i := copy(str, opass)
-	i += copy(str[i:], valSalt)
-	i += copy(str[i:], userKey)
-
-	h := crypt.alg2b(str, opass, userKey)
-
-	O := make([]byte, len(h)+len(valSalt)+len(keySalt))
-	i = copy(O, h[:32])
-	i += copy(O[i:], valSalt)
-	i += copy(O[i:], keySalt)
-
-	crypt.encryptStd.O = O
-
-	// step b: compute OE (owner encryption)
-
-	// str still contains a password and a user key - reuse both, but overwrite the salt
-	i = len(opass)
-	i += copy(str[i:], keySalt)
-	// i += len(userKey)
-
-	h = crypt.alg2b(str, opass, userKey)
-
-	ac, err := aes.NewCipher(h[:32])
-	if err != nil {
-		panic(err)
-	}
-
-	iv := crypt.aesZeroIV()
-	cbc := cipher.NewCBCEncrypter(ac, iv)
-	OE := make([]byte, 32)
-	cbc.CryptBlocks(OE, crypt.encryptionKey[:32])
-	crypt.encryptStd.OE = OE
-
-	return nil
-}
-
-// alg10 computes the encryption dictionary's Perms (permissions) value (R=6).
-// 7.6.4.4.8 Algorithm 10 (page 87)
-func (crypt *PdfCrypt) alg10() error {
-	// step a: extend permissions to 64 bits
-	perms := uint64(uint32(crypt.encryptStd.P)) | (math.MaxUint32 << 32)
-
-	// step b: record permissions
-	Perms := make([]byte, 16)
-	binary.LittleEndian.PutUint64(Perms[:8], perms)
-
-	// step c: record EncryptMetadata
-	if crypt.encryptStd.EncryptMetadata {
-		Perms[8] = 'T'
-	} else {
-		Perms[8] = 'F'
-	}
-
-	// step d: write "adb" magic
-	copy(Perms[9:12], "adb")
-
-	// step e: write 4 bytes of random data
-
-	// spec doesn't specify them as generated "from a strong random source",
-	// but we will use the cryptographic random generator anyway
-	if _, err := io.ReadFull(rand.Reader, Perms[12:16]); err != nil {
-		return err
-	}
-
-	// step f: encrypt permissions
-	ac, err := aes.NewCipher(crypt.encryptionKey[:32])
-	if err != nil {
-		panic(err)
-	}
-
-	ecb := newECBEncrypter(ac)
-	ecb.CryptBlocks(Perms, Perms)
-
-	crypt.encryptStd.Perms = Perms[:16]
-	return nil
-}
-
-// alg11 authenticates the user password (R >= 5) and returns the hash.
-func (crypt *PdfCrypt) alg11(upass []byte) ([]byte, error) {
-	str := make([]byte, len(upass)+8)
-	i := copy(str, upass)
-	i += copy(str[i:], crypt.encryptStd.U[32:40]) // user Validation Salt
-
-	h := crypt.alg2b(str, upass, nil)
-	h = h[:32]
-	if !bytes.Equal(h, crypt.encryptStd.U[:32]) {
-		return nil, nil
-	}
-	return h, nil
-}
-
-// alg12 authenticates the owner password (R >= 5) and returns the hash.
-// 7.6.4.4.10 Algorithm 12 (page 87)
-func (crypt *PdfCrypt) alg12(opass []byte) ([]byte, error) {
-	str := make([]byte, len(opass)+8+48)
-	i := copy(str, opass)
-	i += copy(str[i:], crypt.encryptStd.O[32:40]) // owner Validation Salt
-	i += copy(str[i:], crypt.encryptStd.U[0:48])
-
-	h := crypt.alg2b(str, opass, crypt.encryptStd.U[0:48])
-	h = h[:32]
-	if !bytes.Equal(h, crypt.encryptStd.O[:32]) {
-		return nil, nil
-	}
-	return h, nil
-}
-
-// alg13 validates user permissions (P+EncryptMetadata vs Perms) for R=6.
-// 7.6.4.4.11 Algorithm 13 (page 87)
-func (crypt *PdfCrypt) alg13(fkey []byte) (bool, error) {
-	perms := make([]byte, 16)
-	copy(perms, crypt.encryptStd.Perms[:16])
-
-	ac, err := aes.NewCipher(fkey[:32])
-	if err != nil {
-		panic(err)
-	}
-
-	ecb := newECBDecrypter(ac)
-	ecb.CryptBlocks(perms, perms)
-
-	if !bytes.Equal(perms[9:12], []byte("adb")) {
-		return false, errors.New("decoded permissions are invalid")
-	}
-	p := AccessPermissions(binary.LittleEndian.Uint32(perms[0:4]))
-	if p != crypt.encryptStd.P {
-		return false, errors.New("permissions validation failed")
-	}
-	encMeta := true
-	if perms[8] == 'T' {
-		encMeta = true
-	} else if perms[8] == 'F' {
-		encMeta = false
-	} else {
-		return false, errors.New("decoded metadata encryption flag is invalid")
-	}
-	if encMeta != crypt.encryptStd.EncryptMetadata {
-		return false, errors.New("metadata encryption validation failed")
-	}
-	return true, nil
 }
