+++ pdf/core/crypt_test.go
@@ -8,154 +8,13 @@
 package core
 
 import (
-	"bytes"
-	"fmt"
-	"math"
-	"math/rand"
-	"strings"
 	"testing"
-	"time"
 
 	"github.com/unidoc/unidoc/common"
 )
 
 func init() {
 	common.SetLogger(common.ConsoleLogger{})
-}
-
-func TestPadding(t *testing.T) {
-	crypter := PdfCrypt{}
-
-	// Case 1 empty pass, should match padded string.
-	key := crypter.paddedPass([]byte(""))
-	if len(key) != 32 {
-		t.Errorf("Fail, expected padded pass length = 32 (%d)", len(key))
-	}
-	if key[0] != 0x28 {
-		t.Errorf("key[0] != 0x28 (%q in %q)", key[0], key)
-	}
-	if key[31] != 0x7A {
-		t.Errorf("key[31] != 0x7A (%q in %q)", key[31], key)
-	}
-
-	// Case 2, non empty pass.
-	key = crypter.paddedPass([]byte("bla"))
-	if len(key) != 32 {
-		t.Errorf("Fail, expected padded pass length = 32 (%d)", len(key))
-	}
-	if string(key[0:3]) != "bla" {
-		t.Errorf("Expecting start with bla (%s)", key)
-	}
-	if key[3] != 0x28 {
-		t.Errorf("key[3] != 0x28 (%q in %q)", key[3], key)
-	}
-	if key[31] != 0x64 {
-		t.Errorf("key[31] != 0x64 (%q in %q)", key[31], key)
-	}
-}
-
-// Test algorithm 2.
-func TestAlg2(t *testing.T) {
-	crypter := PdfCrypt{
-		encrypt: encryptDict{
-			V:      2,
-			Length: 128,
-		},
-		encryptStd: stdEncryptDict{
-			R:               3,
-			P:               0xfffff0c0,
-			EncryptMetadata: true,
-			O: []byte{0xE6, 0x00, 0xEC, 0xC2, 0x02, 0x88, 0xAD, 0x8B,
-				0x5C, 0x72, 0x64, 0xA9, 0x5C, 0x29, 0xC6, 0xA8, 0x3E, 0xE2, 0x51,
-				0x76, 0x79, 0xAA, 0x02, 0x18, 0xBE, 0xCE, 0xEA, 0x8B, 0x79, 0x86,
-				0x72, 0x6A, 0x8C, 0xDB},
-		},
-		id0: string([]byte{0x4e, 0x00, 0x99, 0xe5, 0x36, 0x78, 0x93, 0x24,
-			0xff, 0xd5, 0x82, 0xe4, 0xec, 0x0e, 0xa3, 0xb4}),
-	}
-
-	key := crypter.alg2([]byte(""))
-
-	keyExp := []byte{0xf8, 0x94, 0x9c, 0x5a, 0xf5, 0xa0, 0xc0, 0xca,
-		0x30, 0xb8, 0x91, 0xc1, 0xbb, 0x2c, 0x4f, 0xf5}
-
-	if string(key) != string(keyExp) {
-		common.Log.Debug("   Key (%d): % x", len(key), key)
-		common.Log.Debug("KeyExp (%d): % x", len(keyExp), keyExp)
-		t.Errorf("alg2 -> key != expected\n")
-	}
-
-}
-
-// Test algorithm 3.
-func TestAlg3(t *testing.T) {
-	crypter := PdfCrypt{
-		encrypt: encryptDict{
-			V:      2,
-			Length: 128,
-		},
-		encryptStd: stdEncryptDict{
-			R:               3,
-			P:               0xfffff0c0,
-			EncryptMetadata: true,
-		},
-		id0: string([]byte{0x4e, 0x00, 0x99, 0xe5, 0x36, 0x78, 0x93, 0x24,
-			0xff, 0xd5, 0x82, 0xe4, 0xec, 0x0e, 0xa3, 0xb4}),
-	}
-
-	Oexp := []byte{0xE6, 0x00, 0xEC, 0xC2, 0x02, 0x88, 0xAD, 0x8B,
-		0x0d, 0x64, 0xA9, 0x29, 0xC6, 0xA8, 0x3E, 0xE2, 0x51,
-		0x76, 0x79, 0xAA, 0x02, 0x18, 0xBE, 0xCE, 0xEA, 0x8B, 0x79, 0x86,
-		0x72, 0x6A, 0x8C, 0xDB}
-	O, err := crypter.alg3([]byte(""), []byte("test"))
-	if err != nil {
-		t.Errorf("crypt alg3 error %s", err)
-		return
-	}
-
-	if string(O) != string(Oexp) {
-		common.Log.Debug("   O (%d): % x", len(O), O)
-		common.Log.Debug("Oexp (%d): % x", len(Oexp), Oexp)
-		t.Errorf("alg3 -> key != expected")
-	}
-}
-
-// Test algorithm 5 for computing dictionary's U (user password) value
-// valid for R >= 3.
-func TestAlg5(t *testing.T) {
-	crypter := PdfCrypt{
-		encrypt: encryptDict{
-			V:      2,
-			Length: 128,
-		},
-		encryptStd: stdEncryptDict{
-			R:               3,
-			P:               0xfffff0c0,
-			EncryptMetadata: true,
-			O: []byte{0xE6, 0x00, 0xEC, 0xC2, 0x02, 0x88, 0xAD, 0x8B,
-				0x5C, 0x72, 0x64, 0xA9, 0x5C, 0x29, 0xC6, 0xA8, 0x3E, 0xE2, 0x51,
-				0x76, 0x79, 0xAA, 0x02, 0x18, 0xBE, 0xCE, 0xEA, 0x8B, 0x79, 0x86,
-				0x72, 0x6A, 0x8C, 0xDB},
-		},
-		id0: string([]byte{0x4e, 0x00, 0x99, 0xe5, 0x36, 0x78, 0x93, 0x24,
-			0xff, 0xd5, 0x82, 0xe4, 0xec, 0x0e, 0xa3, 0xb4}),
-	}
-
-	U, _, err := crypter.alg5([]byte(""))
-	if err != nil {
-		t.Errorf("Error %s", err)
-		return
-	}
-
-	Uexp := []byte{0x59, 0x66, 0x38, 0x6c, 0x76, 0xfe, 0x95, 0x7d, 0x3d,
-		0x0d, 0x14, 0x3d, 0x36, 0xfd, 0x01, 0x3d, 0x00, 0x00, 0x00, 0x00,
-		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00}
-
-	if string(U[0:16]) != string(Uexp[0:16]) {
-		common.Log.Info("   U (%d): % x", len(U), U)
-		common.Log.Info("Uexp (%d): % x", len(Uexp), Uexp)
-		t.Errorf("U != expected\n")
-	}
 }
 
 // Test decrypting. Example with V=2, R=3, using standard algorithm.
@@ -233,123 +92,5 @@ func TestDecryption1(t *testing.T) {
 	if string(so.Stream) != string(exp) {
 		t.Errorf("Stream content wrong")
 		return
-	}
-}
-
-func BenchmarkAlg2b(b *testing.B) {
-	// hash runs a variable number of rounds, so we need to have a
-	// deterministic random source to make benchmark results comparable
-	r := rand.New(rand.NewSource(1234567))
-	const n = 20
-	pass := make([]byte, n)
-	r.Read(pass)
-	data := make([]byte, n+8+48)
-	r.Read(data)
-	user := make([]byte, 48)
-	r.Read(user)
-
-	b.ResetTimer()
-	b.ReportAllocs()
-	for i := 0; i < b.N; i++ {
-		_ = alg2b(data, pass, user)
-	}
-}
-
-func TestAESv3(t *testing.T) {
-	const keySize = 32
-
-	seed := time.Now().UnixNano()
-	rand := rand.New(rand.NewSource(seed))
-
-	var cases = []struct {
-		Name      string
-		EncMeta   bool
-		UserPass  string
-		OwnerPass string
-	}{
-		{
-			Name: "simple", EncMeta: true,
-			UserPass: "user", OwnerPass: "owner",
-		},
-		{
-			Name: "utf8", EncMeta: false,
-			UserPass: "æøå-u", OwnerPass: "æøå-o",
-		},
-		{
-			Name: "long", EncMeta: true,
-			UserPass:  strings.Repeat("user", 80),
-			OwnerPass: strings.Repeat("owner", 80),
-		},
-	}
-
-	const (
-		perms = 0x12345678
-	)
-
-	for _, R := range []int{5, 6} {
-		R := R
-		t.Run(fmt.Sprintf("R=%d", R), func(t *testing.T) {
-			for _, c := range cases {
-				c := c
-				t.Run(c.Name, func(t *testing.T) {
-					fkey := make([]byte, keySize)
-					rand.Read(fkey)
-
-					crypt := &PdfCrypt{
-						encrypt: encryptDict{
-							V: 5,
-						},
-						encryptStd: stdEncryptDict{
-							R: R, P: perms,
-							EncryptMetadata: c.EncMeta,
-						},
-						encryptionKey: append([]byte{}, fkey...),
-					}
-
-					// generate encryption parameters
-					err := crypt.generateR6([]byte(c.UserPass), []byte(c.OwnerPass))
-					if err != nil {
-						t.Fatal("Failed to encrypt:", err)
-					}
-
-					// Perms and EncryptMetadata are checked as a part of alg2a
-
-					// decrypt using user password
-					crypt.encryptionKey = nil
-					ok, err := crypt.alg2a([]byte(c.UserPass))
-					if err != nil || !ok {
-						t.Error("Failed to authenticate user pass:", err)
-					} else if !bytes.Equal(crypt.encryptionKey, fkey) {
-						t.Error("wrong encryption key")
-					}
-
-					// decrypt using owner password
-					crypt.encryptionKey = nil
-					ok, err = crypt.alg2a([]byte(c.OwnerPass))
-					if err != nil || !ok {
-						t.Error("Failed to authenticate owner pass:", err)
-					} else if !bytes.Equal(crypt.encryptionKey, fkey) {
-						t.Error("wrong encryption key")
-					}
-
-					// try to elevate user permissions
-					crypt.encryptStd.P = math.MaxUint32
-
-					crypt.encryptionKey = nil
-					ok, err = crypt.alg2a([]byte(c.UserPass))
-					if R == 5 {
-						// it's actually possible with R=5, since Perms is not generated
-						if err != nil || !ok {
-							t.Error("Failed to authenticate user pass:", err)
-						}
-					} else {
-						// not possible in R=6, should return an error
-						if err == nil || ok {
-							t.Error("was able to elevate permissions with R=6")
-						}
-					}
-				})
-			}
-		})
 	}
 }
